API Rename classes to be less confusing with the namespaces
