+++ .upgrade.yml
@@ -1,10 +1,10 @@
 mappings:
-  GridFieldBulkActionDeleteHandler: Colymba\BulkManager\GridFieldBulkActionDeleteHandler
-  GridFieldBulkActionEditHandler: Colymba\BulkManager\GridFieldBulkActionEditHandler
-  GridFieldBulkActionHandler: Colymba\BulkManager\GridFieldBulkActionHandler
-  GridFieldBulkActionUnlinkHandler: Colymba\BulkManager\GridFieldBulkActionUnlinkHandler
-  GridFieldBulkManager: Colymba\BulkManager\GridFieldBulkManager
+  GridFieldBulkActionDeleteHandler: Colymba\BulkManager\BulkAction\DeleteHandler
+  GridFieldBulkActionEditHandler: Colymba\BulkManager\BulkAction\EditHandler
+  GridFieldBulkActionHandler: Colymba\BulkManager\BulkAction\Handler
+  GridFieldBulkActionUnlinkHandler: Colymba\BulkManager\BulkAction\UnlinkHandler
+  GridFieldBulkManager: Colymba\BulkManager\BulkManager
   BulkUploadField: Colymba\BulkUpload\BulkUploadField
   GridFieldBulkImageUpload: Colymba\BulkUpload\GridFieldBulkImageUpload
-  GridFieldBulkUpload: Colymba\BulkUpload\GridFieldBulkUpload
-  GridFieldBulkUpload_Request: Colymba\BulkUpload\GridFieldBulkUpload_Request
+  GridFieldBulkUpload: Colymba\BulkUpload\BulkUploader
+  GridFieldBulkUpload_Request: Colymba\BulkUpload\BulkUploadRequest
