+++ bulkUpload/code/GridFieldBulkImageUpload.php
@@ -2,22 +2,22 @@
 
 namespace Colymba\BulkUpload;
 
-use Colymba\BulkUpload\GridFieldBulkUpload;
+use Colymba\BulkUpload\BulkUploader;
 use SilverStripe\Dev\Deprecation;
 
 /**
  * Legacy GridFieldBulkImageUpload component.
  *
- * @deprecated 2.0 "GridFieldBulkImageUpload" is deprecated, use {@link GridFieldBulkUpload} class instead.
+ * @deprecated 2.0 "GridFieldBulkImageUpload" is deprecated, use {@link BulkUploader} class instead.
  *
  * @author colymba
  */
-class GridFieldBulkImageUpload extends GridFieldBulkUpload
+class GridFieldBulkImageUpload extends BulkUploader
 {
     /**
      * Component constructor.
      *
-     * @deprecated 2.0 "GridFieldBulkImageUpload" is deprecated, use {@link GridFieldBulkUpload} class instead.
+     * @deprecated 2.0 "GridFieldBulkImageUpload" is deprecated, use {@link BulkUploader} class instead.
      *
      * @param string $fileRelationName
      */
@@ -25,9 +25,9 @@ class GridFieldBulkImageUpload extends GridFieldBulkUpload
     {
         Deprecation::notice(
             '2.0',
-            '"GridFieldBulkImageUpload" is deprecated, use "GridFieldBulkUpload" class instead.'
+            '"GridFieldBulkImageUpload" is deprecated, use "BulkUploader" class instead.'
         );
 
-        return new GridFieldBulkUpload($fileRelationName);
+        return new BulkUploader($fileRelationName);
     }
 }
